+++ plugin-name/Admin/PluginNameAdmin.php
@@ -1,26 +1,26 @@
-<?php namespace Pub;
+<?php namespace PluginName\Admin;
 
 /**
- * The public-facing functionality of the plugin.
+ * The admin-specific functionality of the plugin.
  *
  * @link       http://example.com
  * @since      1.0.0
  *
  * @package    PluginName
- * @subpackage PluginName/public
+ * @subpackage PluginName/admin
  */
 
 /**
- * The public-facing functionality of the plugin.
+ * The admin-specific functionality of the plugin.
  *
  * Defines the plugin name, version, and two examples hooks for how to
  * enqueue the admin-specific stylesheet and JavaScript.
  *
  * @package    PluginName
- * @subpackage PluginName/public
+ * @subpackage PluginName/admin
  * @author     Your Name <email@example.com>
  */
-class PluginNamePublic {
+class PluginNameAdmin {
 
 	/**
 	 * The ID of this plugin.
@@ -44,7 +44,7 @@ class PluginNamePublic {
 	 * Initialize the class and set its properties.
 	 *
 	 * @since    1.0.0
-	 * @param      string    $pluginName       The name of the plugin.
+	 * @param      string    $pluginName       The name of this plugin.
 	 * @param      string    $version    The version of this plugin.
 	 */
 	public function __construct( $pluginName, $version ) {
@@ -55,7 +55,7 @@ class PluginNamePublic {
 	}
 
 	/**
-	 * Register the stylesheets for the public-facing side of the site.
+	 * Register the stylesheets for the admin area.
 	 *
 	 * @since    1.0.0
 	 */
@@ -73,12 +73,12 @@ class PluginNamePublic {
 		 * class.
 		 */
 
-		wp_enqueue_style( $this->pluginName, plugin_dir_url( __FILE__ ) . 'css/plugin-name-public.css', array(), $this->version, 'all' );
+		wp_enqueue_style( $this->pluginName, plugin_dir_url( __FILE__ ) . 'css/plugin-name-admin.css', array(), $this->version, 'all' );
 
 	}
 
 	/**
-	 * Register the stylesheets for the public-facing side of the site.
+	 * Register the JavaScript for the admin area.
 	 *
 	 * @since    1.0.0
 	 */
@@ -96,7 +96,7 @@ class PluginNamePublic {
 		 * class.
 		 */
 
-		wp_enqueue_script( $this->pluginName, plugin_dir_url( __FILE__ ) . 'js/plugin-name-public.js', array( 'jquery' ), $this->version, false );
+		wp_enqueue_script( $this->pluginName, plugin_dir_url( __FILE__ ) . 'js/plugin-name-admin.js', array( 'jquery' ), $this->version, false );
 
 	}
