+++ plugin-name/Lib/PluginNameActivator.php
@@ -1,7 +1,7 @@
-<?php namespace Lib;
+<?php namespace PluginName\Lib;
 
 /**
- * Fired during plugin deactivation
+ * Fired during plugin activation
  *
  * @link       http://example.com
  * @since      1.0.0
@@ -11,16 +11,16 @@
  */
 
 /**
- * Fired during plugin deactivation.
+ * Fired during plugin activation.
  *
- * This class defines all code necessary to run during the plugin's deactivation.
+ * This class defines all code necessary to run during the plugin's activation.
  *
  * @since      1.0.0
  * @package    PluginName
  * @subpackage PluginName/includes
  * @author     Your Name <email@example.com>
  */
-class PluginNameDeactivator {
+class PluginNameActivator {
 
 	/**
 	 * Short Description. (use period)
@@ -29,7 +29,7 @@ class PluginNameDeactivator {
 	 *
 	 * @since    1.0.0
 	 */
-	public static function deactivate() {
+	public static function activate() {
 
 	}
