+++ plugin-name/Public/css/plugin-name-public.css
