+++ plugin-name/Public/js/plugin-name-public.js
